@@ -1,67 +1,66 @@
 {
     "name": "hansadema/discoverify",
-    "description": "Discover music on Spotify using curated YouTube music channels.",
-    "keywords": ["youtube", "spotify"],
-    "license": "MIT",
     "type": "project",
+    "description": "Discover music on Spotify using curated YouTube music channels.",
+    "keywords": [
+        "youtube",
+        "spotify"
+    ],
+    "license": "MIT",
     "require": {
-        "php": "^7.1.3",
-        "ext-simplexml": "*",
-        "doctrine/inflector": "1.1.*",
+        "php": "^7.2",
         "fideloper/proxy": "^4.0",
         "guzzlehttp/guzzle": "~6.0",
         "jwilsson/spotify-web-api-php": "^2.6",
-        "laravel/framework": "5.7.*",
-        "laravel/socialite": "^3.0",
-        "laravel/tinker": "^1.0",
-        "predis/predis": "^1.1",
+        "laravel/framework": "^6.2",
+        "laravel/socialite": "^4.0",
+        "laravel/tinker": "^2.0",
+        "sentry/sentry-laravel": "^1.5",
         "socialiteproviders/spotify": "^3.0"
     },
     "require-dev": {
-        "beyondcode/laravel-dump-server": "^1.0",
-        "filp/whoops": "^2.0",
+        "facade/ignition": "^1.4",
         "fzaninotto/faker": "^1.4",
         "mockery/mockery": "^1.0",
-        "nunomaduro/collision": "^2.0",
+        "nunomaduro/collision": "^3.0",
         "phpunit/phpunit": "^7.0"
     },
+    "config": {
+        "optimize-autoloader": true,
+        "preferred-install": "dist",
+        "sort-packages": true
+    },
+    "extra": {
+        "laravel": {
+            "dont-discover": []
+        }
+    },
     "autoload": {
+        "psr-4": {
+            "App\\": "app/"
+        },
         "classmap": [
             "database/seeds",
             "database/factories"
-        ],
-        "psr-4": {
-            "App\\": "app/"
-        }
+        ]
     },
     "autoload-dev": {
         "psr-4": {
             "Tests\\": "tests/"
         }
     },
-    "extra": {
-        "laravel": {
-            "dont-discover": [
-            ]
-        }
-    },
+    "minimum-stability": "dev",
+    "prefer-stable": true,
     "scripts": {
+        "post-autoload-dump": [
+            "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
+            "@php artisan package:discover --ansi"
+        ],
         "post-root-package-install": [
             "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
         ],
         "post-create-project-cmd": [
-            "@php artisan key:generate"
-        ],
-        "post-autoload-dump": [
-            "Illuminate\\Foundation\\ComposerScripts::postAutoloadDump",
-            "@php artisan package:discover"
+            "@php artisan key:generate --ansi"
         ]
-    },
-    "config": {
-        "preferred-install": "dist",
-        "sort-packages": true,
-        "optimize-autoloader": true
-    },
-    "minimum-stability": "dev",
-    "prefer-stable": true
+    }
 }
