@@ -10,10 +10,12 @@ class Authenticate extends Middleware
      * Get the path the user should be redirected to when they are not authenticated.
      *
      * @param  \Illuminate\Http\Request  $request
-     * @return string
+     * @return string|null
      */
     protected function redirectTo($request)
     {
-        return route('login');
+        if (! $request->expectsJson()) {
+            return route('login');
+        }
     }
 }
