@@ -1,18 +1,6 @@
 <?php
 
-if (!env('DB_DATABASE')) {
-    if (env('DATABASE_URL')) {
-        $url = parse_url(env('DATABASE_URL'));
-    } elseif (env('JAWSDB_URL')) {
-        $url = parse_url(env('JAWSDB_URL'));
-    }
-}
-
-if (env('REDIS_URL')) {
-    $redisUrl = parse_url(env('REDIS_URL'));
-} elseif (env('REDISCLOUD_URL')) {
-    $redisUrl = parse_url(env('REDISCLOUD_URL'));
-}
+use Illuminate\Support\Str;
 
 return [
 
@@ -49,27 +37,35 @@ return [
 
         'sqlite' => [
             'driver' => 'sqlite',
+            'url' => env('DATABASE_URL'),
             'database' => env('DB_DATABASE', database_path('database.sqlite')),
             'prefix' => '',
+            'foreign_key_constraints' => env('DB_FOREIGN_KEYS', true),
         ],
 
         'mysql' => [
             'driver' => 'mysql',
-            'host' => env('DB_HOST', isset($url) ? $url['host'] : '127.0.0.1'),
-            'port' => env('DB_PORT', isset($url) ? $url['port'] : '3306'),
-            'database' => env('DB_DATABASE', isset($url) ? ltrim($url['path'], '/') : 'forge'),
-            'username' => env('DB_USERNAME', isset($url) ? $url['user'] : 'forge'),
-            'password' => env('DB_PASSWORD', isset($url) ? $url['pass'] : ''),
+            'url' => env('DATABASE_URL'),
+            'host' => env('DB_HOST', '127.0.0.1'),
+            'port' => env('DB_PORT', '3306'),
+            'database' => env('DB_DATABASE', 'forge'),
+            'username' => env('DB_USERNAME', 'forge'),
+            'password' => env('DB_PASSWORD', ''),
             'unix_socket' => env('DB_SOCKET', ''),
             'charset' => 'utf8mb4',
             'collation' => 'utf8mb4_unicode_ci',
             'prefix' => '',
+            'prefix_indexes' => true,
             'strict' => true,
             'engine' => null,
+            'options' => extension_loaded('pdo_mysql') ? array_filter([
+                PDO::MYSQL_ATTR_SSL_CA => env('MYSQL_ATTR_SSL_CA'),
+            ]) : [],
         ],
 
         'pgsql' => [
             'driver' => 'pgsql',
+            'url' => env('DATABASE_URL'),
             'host' => env('DB_HOST', '127.0.0.1'),
             'port' => env('DB_PORT', '5432'),
             'database' => env('DB_DATABASE', 'forge'),
@@ -77,12 +73,14 @@ return [
             'password' => env('DB_PASSWORD', ''),
             'charset' => 'utf8',
             'prefix' => '',
+            'prefix_indexes' => true,
             'schema' => 'public',
             'sslmode' => 'prefer',
         ],
 
         'sqlsrv' => [
             'driver' => 'sqlsrv',
+            'url' => env('DATABASE_URL'),
             'host' => env('DB_HOST', 'localhost'),
             'port' => env('DB_PORT', '1433'),
             'database' => env('DB_DATABASE', 'forge'),
@@ -90,6 +88,7 @@ return [
             'password' => env('DB_PASSWORD', ''),
             'charset' => 'utf8',
             'prefix' => '',
+            'prefix_indexes' => true,
         ],
 
     ],
@@ -113,27 +112,34 @@ return [
     |--------------------------------------------------------------------------
     |
     | Redis is an open source, fast, and advanced key-value store that also
-    | provides a richer set of commands than a typical key-value systems
+    | provides a richer body of commands than a typical key-value system
     | such as APC or Memcached. Laravel makes it easy to dig right in.
     |
     */
 
     'redis' => [
 
-        'client' => 'predis',
+        'client' => env('REDIS_CLIENT', 'phpredis'),
+
+        'options' => [
+            'cluster' => env('REDIS_CLUSTER', 'redis'),
+            'prefix' => env('REDIS_PREFIX', Str::slug(env('APP_NAME', 'laravel'), '_').'_database_'),
+        ],
 
         'default' => [
-            'host' => env('REDIS_HOST', isset($redisUrl) ? $redisUrl['host'] : '127.0.0.1'),
-            'password' => env('REDIS_PASSWORD', isset($redisUrl) ? $redisUrl['pass'] : null),
-            'port' => env('REDIS_PORT', isset($redisUrl) ? $redisUrl['port'] : 6379),
-            'database' => env('REDIS_DB', 0),
+            'url' => env('REDIS_URL'),
+            'host' => env('REDIS_HOST', '127.0.0.1'),
+            'password' => env('REDIS_PASSWORD', null),
+            'port' => env('REDIS_PORT', '6379'),
+            'database' => env('REDIS_DB', '0'),
         ],
 
         'cache' => [
-            'host' => env('REDIS_HOST', isset($redisUrl) ? $redisUrl['host'] : '127.0.0.1'),
-            'password' => env('REDIS_PASSWORD', isset($redisUrl) ? $redisUrl['pass'] : null),
-            'port' => env('REDIS_PORT', isset($redisUrl) ? $redisUrl['port'] : 6379),
-            'database' => env('REDIS_CACHE_DB', 1),
+            'url' => env('REDIS_URL'),
+            'host' => env('REDIS_HOST', '127.0.0.1'),
+            'password' => env('REDIS_PASSWORD', null),
+            'port' => env('REDIS_PORT', '6379'),
+            'database' => env('REDIS_CACHE_DB', '1'),
         ],
 
     ],
